Check file existence.

+++ cp_to_admin.sh
@@ -62,7 +62,11 @@ if [ ! -d "$LOCAL_ADMIN_PATH" ] ; then
 	sudo cp -a ./system_conf/* $LOCAL_CONF_PATH
 	sudo cp -a ./*.sh $LOCAL_ADMIN_PATH
 	sudo chown -R root:root $LOCAL_ADMIN_PATH
-	sudo ln -s $PWD/config/home-bin/ ~/bin
-	#sudo chown $USER:$USER ~/bin
-	sudo ln -s $PWD/setting ~/setting
+	if [ ! -e "~/bin" ] ; then
+		sudo ln -s $PWD/config/home-bin/ ~/bin
+		#sudo chown $USER:$USER ~/bin
+	fi
+	if [ ! -e "~/setting" ] ; then
+		sudo ln -s $PWD/setting ~/setting
+	fi
 fi
